Use u32 for command offsets in DB
rusqlite@0.38 removes support for usize in the DB, because sqlite doesn't support unsigned 64-bit integers. We don't need offsets nearly that big though because the max history length is 100 commands. This will prep us for the upgrade when it becomes possible (once rusqlite_migration 2.4.0 is release).

+++ crates/core/src/database.rs
@@ -28,7 +28,7 @@ use uuid::Uuid;
 
 /// Maximum number of commands to store in history **per collection**. When we
 /// hit the cap, the oldest commands get evicted.
-const MAX_COMMAND_HISTORY_SIZE: usize = 100;
+const MAX_COMMAND_HISTORY_SIZE: u32 = 100;
 
 /// A SQLite database for persisting data. Generally speaking, any error that
 /// occurs *after* opening the DB connection should be an internal bug, but
@@ -838,7 +838,7 @@ impl CollectionDatabase {
     ///   prompt excludes the current text from results to prevent duplicates.
     pub fn get_command(
         &self,
-        offset: usize,
+        offset: u32,
         exclude: &str,
     ) -> Result<Option<String>, DatabaseError> {
         trace!(offset, "Getting command from history with offset");
